+++ library/Zend/Filter/Compress/AbstractCompressionAdapter.php
@@ -16,4 +16,32 @@ use Zend\Stdlib\AbstractOptions;
  */
 abstract class AbstractCompressionAdapter implements CompressionAdapterInterface
 {
+    /**
+     * @param array $options
+     */
+    public function __construct(array $options = array())
+    {
+        $this->setOptions($options);
+    }
+
+    /**
+     * @TODO: replace with a generic trait that can be reused
+     *
+     * Set options for the given filter
+     *
+     * This method inflect the key names and call the corresponding
+     *
+     * @param array $options
+     */
+    public function setOptions(array $options)
+    {
+        foreach ($options as $key => $value) {
+            $setter = 'set' . str_replace('_', '', $key);
+
+            if (method_exists($this, $setter)) {
+                $this->$setter($value);
+                continue;
+            }
+        }
+    }
 }
