+++ tests/ZendTest/Crypt/Filter/EncryptTest.php
@@ -8,9 +8,9 @@
  * @package   Zend_Filter
  */
 
-namespace ZendTest\Filter;
+namespace ZendTest\Crypt\Filter;
 
-use Zend\Filter\Encrypt as EncryptFilter;
+use Zend\Crypt\Filter\Encrypt as EncryptFilter;
 
 /**
  * @category   Zend
