+++ tests/ZendTest/Crypt/Filter/File/EncryptTest.php
@@ -10,8 +10,8 @@
 
 namespace ZendTest\Filter\File;
 
-use Zend\Filter\File\Decrypt as FileDecrypt;
-use Zend\Filter\File\Encrypt as FileEncrypt;
+use Zend\Crypt\Filter\File\Encrypt as FileEncrypt;
+use Zend\Crypt\Filter\File\Decrypt as FileDecrypt;
 
 /**
  * @category   Zend
@@ -19,7 +19,7 @@ use Zend\Filter\File\Encrypt as FileEncrypt;
  * @subpackage UnitTests
  * @group      Zend_Filter
  */
-class DecryptTest extends \PHPUnit_Framework_TestCase
+class EncryptTest extends \PHPUnit_Framework_TestCase
 {
     public function setUp()
     {
@@ -30,20 +30,12 @@ class DecryptTest extends \PHPUnit_Framework_TestCase
         if (file_exists(dirname(__DIR__).'/_files/newencryption.txt')) {
             unlink(dirname(__DIR__).'/_files/newencryption.txt');
         }
-
-        if (file_exists(dirname(__DIR__).'/_files/newencryption2.txt')) {
-            unlink(dirname(__DIR__).'/_files/newencryption2.txt');
-        }
     }
 
     public function tearDown()
     {
         if (file_exists(dirname(__DIR__).'/_files/newencryption.txt')) {
             unlink(dirname(__DIR__).'/_files/newencryption.txt');
-        }
-
-        if (file_exists(dirname(__DIR__).'/_files/newencryption2.txt')) {
-            unlink(dirname(__DIR__).'/_files/newencryption2.txt');
         }
     }
 
@@ -62,22 +54,16 @@ class DecryptTest extends \PHPUnit_Framework_TestCase
             $filter->getFilename());
 
         $filter->setKey('1234567890123456');
-        $filter->filter(dirname(__DIR__).'/_files/encryption.txt');
+        $this->assertEquals(dirname(__DIR__).'/_files/newencryption.txt',
+            $filter->filter(dirname(__DIR__).'/_files/encryption.txt'));
 
-        $filter = new FileDecrypt();
+        $this->assertEquals(
+            'Encryption',
+            file_get_contents(dirname(__DIR__).'/_files/encryption.txt'));
 
         $this->assertNotEquals(
             'Encryption',
             file_get_contents(dirname(__DIR__).'/_files/newencryption.txt'));
-
-        $filter->setKey('1234567890123456');
-        $this->assertEquals(
-            dirname(__DIR__).'/_files/newencryption.txt',
-            $filter->filter(dirname(__DIR__).'/_files/newencryption.txt'));
-
-        $this->assertEquals(
-            'Encryption',
-            trim(file_get_contents(dirname(__DIR__).'/_files/newencryption.txt')));
     }
 
     public function testEncryptionWithDecryption()
@@ -93,19 +79,13 @@ class DecryptTest extends \PHPUnit_Framework_TestCase
             file_get_contents(dirname(__DIR__).'/_files/newencryption.txt'));
 
         $filter = new FileDecrypt();
-        $filter->setFilename(dirname(__DIR__).'/_files/newencryption2.txt');
-
-        $this->assertEquals(
-            dirname(__DIR__).'/_files/newencryption2.txt',
-            $filter->getFilename());
-
         $filter->setKey('1234567890123456');
         $input = $filter->filter(dirname(__DIR__).'/_files/newencryption.txt');
-        $this->assertEquals(dirname(__DIR__).'/_files/newencryption2.txt', $input);
+        $this->assertEquals(dirname(__DIR__).'/_files/newencryption.txt', $input);
 
         $this->assertEquals(
             'Encryption',
-            trim(file_get_contents(dirname(__DIR__).'/_files/newencryption2.txt')));
+            trim(file_get_contents(dirname(__DIR__).'/_files/newencryption.txt')));
     }
 
     /**
@@ -113,10 +93,26 @@ class DecryptTest extends \PHPUnit_Framework_TestCase
      */
     public function testNonExistingFile()
     {
-        $filter = new FileDecrypt();
-        $filter->setVector('1234567890123456');
+        $filter = new FileEncrypt();
+        $filter->setKey('1234567890123456');
 
         $this->setExpectedException('\Zend\Filter\Exception\InvalidArgumentException', 'not found');
-        $filter->filter(dirname(__DIR__).'/_files/nofile.txt');
+        echo $filter->filter(dirname(__DIR__).'/_files/nofile.txt');
+    }
+
+    /**
+     * @return void
+     */
+    public function testEncryptionInSameFile()
+    {
+        $filter = new FileEncrypt();
+        $filter->setKey('1234567890123456');
+
+        copy(dirname(__DIR__).'/_files/encryption.txt', dirname(__DIR__).'/_files/newencryption.txt');
+        $filter->filter(dirname(__DIR__).'/_files/newencryption.txt');
+
+        $this->assertNotEquals(
+            'Encryption',
+            trim(file_get_contents(dirname(__DIR__).'/_files/newencryption.txt')));
     }
 }
