+++ tests/ZendTest/Filter/StringToLowerTest.php
@@ -23,7 +23,7 @@ class StringToLowerTest extends \PHPUnit_Framework_TestCase
     /**
      * Zend_Filter_StringToLower object
      *
-     * @var Zend_Filter_StringToLower
+     * @var StringToLowerFilter
      */
     protected $_filter;
