+++ tests/ZendTest/Filter/Word/SeparatorToSeparatorTest.php
@@ -35,7 +35,9 @@ class SeparatorToSeparatorTest extends \PHPUnit_Framework_TestCase
     public function testFilterSeparatesWordsWithSearchSpecified()
     {
         $string   = 'dash=separated=words';
-        $filter   = new SeparatorToSeparatorFilter('=');
+        $filter   = new SeparatorToSeparatorFilter(array(
+            'search_separator' => '='
+        ));
         $filtered = $filter($string);
 
         $this->assertNotEquals($string, $filtered);
@@ -45,7 +47,10 @@ class SeparatorToSeparatorTest extends \PHPUnit_Framework_TestCase
     public function testFilterSeparatesWordsWithSearchAndReplacementSpecified()
     {
         $string   = 'dash=separated=words';
-        $filter   = new SeparatorToSeparatorFilter('=', '?');
+        $filter   = new SeparatorToSeparatorFilter(array(
+            'search_separator' => '=',
+            'replacement_separator' => '?'
+        ));
         $filtered = $filter($string);
 
         $this->assertNotEquals($string, $filtered);
